REL: Rearchitecture for alternate operating systems.
Reworked the architecture of the volume interrogator to allow for the support of alternate operating systems. Still limited to darwin (OSX/macOS) operating systems at this time.
